Реализовал ОЗУ

+++ db/Classes/RAM.cs
@@ -16,7 +16,7 @@ namespace db.Classes
         public string typeMemory { get; set; }
 
         [DisplayName("Частота")]
-        public float frequency { get; set; }
+        public int frequency { get; set; }
 
         [DisplayName("Объем")]
         public int volune { get; set; }
@@ -32,7 +32,7 @@ namespace db.Classes
             id = Guid.NewGuid().ToString();
         }
 
-        public RAM(string typeMemory, float frequency, int volune, string title, string manufacturer)
+        public RAM(string typeMemory, int frequency, int volune, string title, string manufacturer)
         {
             id = Guid.NewGuid().ToString();
             this.typeMemory = typeMemory;
